pyproject.toml
[project]
name = "systech-aidd-test"
version = "0.1.0"
description = "AI-driven Telegram bot for testing LLM integration"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.0",
    "openai>=1.0",
    "python-dotenv>=1.0",
    "sqlalchemy[asyncio]>=2.0",
    "aiosqlite>=0.19",
    "alembic>=1.13",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "W",   # pycodestyle warnings
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
    "RUF", # Ruff-specific rules
]
ignore = [
    "RUF001",  # ambiguous characters (false positives for non-Latin text)
    "RUF002",  # ambiguous characters in docstrings
    "RUF003",  # ambiguous characters in comments
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]  # allow assert in tests

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_any_unimported = false
mypy_path = "src"
explicit_package_bases = true
namespace_packages = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]
addopts = "--strict-markers --cov-config=pyproject.toml"

[tool.coverage.run]
omit = [
    "src/main.py",  # Entry point - tested via integration tests
]
